requirements.txt
django==4.2.4
requests==2.31.0
pyTelegramBotAPI==4.13.0
gunicorn==21.2.0
gevent==23.9.1
google-auth-oauthlib==1.2.0
django-environ==0.11.2
djangorestframework==3.15.2
attrs==23.2.0
social-auth-app-django==5.4.1
python-dotenv==1.0.1
pillow==10.4.0
django-countries==7.6.1
django-storages==1.14.4
boto3==1.35.7
django-storages==1.14.4
psycopg2==2.9.9